+++ internal/service/remotecfg/remotecfg_test.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"connectrpc.com/connect"
-	agentv1 "github.com/grafana/agent-remote-config/api/gen/proto/go/agent/v1"
+	collectorv1 "github.com/grafana/alloy-remote-config/api/gen/proto/go/collector/v1"
 	"github.com/grafana/alloy/internal/alloy"
 	"github.com/grafana/alloy/internal/alloy/componenttest"
 	"github.com/grafana/alloy/internal/alloy/logging"
@@ -39,7 +39,7 @@ func TestOnDiskCache(t *testing.T) {
 		url = "%s"
 	`, url)))
 
-	client := &agentClient{}
+	client := &collectorClient{}
 	env.svc.asClient = client
 
 	// Mock client to return an unparseable response.
@@ -73,7 +73,7 @@ func TestAPIResponse(t *testing.T) {
 		poll_frequency = "10ms"
 	`, url)))
 
-	client := &agentClient{}
+	client := &collectorClient{}
 	env.svc.asClient = client
 
 	// Mock client to return a valid response.
@@ -103,10 +103,10 @@ func TestAPIResponse(t *testing.T) {
 	}, time.Second, 10*time.Millisecond)
 }
 
-func buildGetConfigHandler(in string) func(context.Context, *connect.Request[agentv1.GetConfigRequest]) (*connect.Response[agentv1.GetConfigResponse], error) {
-	return func(context.Context, *connect.Request[agentv1.GetConfigRequest]) (*connect.Response[agentv1.GetConfigResponse], error) {
-		rsp := &connect.Response[agentv1.GetConfigResponse]{
-			Msg: &agentv1.GetConfigResponse{
+func buildGetConfigHandler(in string) func(context.Context, *connect.Request[collectorv1.GetConfigRequest]) (*connect.Response[collectorv1.GetConfigResponse], error) {
+	return func(context.Context, *connect.Request[collectorv1.GetConfigRequest]) (*connect.Response[collectorv1.GetConfigResponse], error) {
+		rsp := &connect.Response[collectorv1.GetConfigResponse]{
+			Msg: &collectorv1.GetConfigResponse{
 				Content: in,
 			},
 		}
@@ -179,12 +179,12 @@ func (f fakeHost) NewController(id string) service.Controller {
 	return serviceController{ctrl}
 }
 
-type agentClient struct {
+type collectorClient struct {
 	mut           sync.RWMutex
-	getConfigFunc func(context.Context, *connect.Request[agentv1.GetConfigRequest]) (*connect.Response[agentv1.GetConfigResponse], error)
+	getConfigFunc func(context.Context, *connect.Request[collectorv1.GetConfigRequest]) (*connect.Response[collectorv1.GetConfigResponse], error)
 }
 
-func (ag *agentClient) GetConfig(ctx context.Context, req *connect.Request[agentv1.GetConfigRequest]) (*connect.Response[agentv1.GetConfigResponse], error) {
+func (ag *collectorClient) GetConfig(ctx context.Context, req *connect.Request[collectorv1.GetConfigRequest]) (*connect.Response[collectorv1.GetConfigResponse], error) {
 	ag.mut.RLock()
 	defer ag.mut.RUnlock()
 
@@ -194,19 +194,19 @@ func (ag *agentClient) GetConfig(ctx context.Context, req *connect.Request[agent
 
 	panic("getConfigFunc not set")
 }
-func (ag *agentClient) GetAgent(context.Context, *connect.Request[agentv1.GetAgentRequest]) (*connect.Response[agentv1.Agent], error) {
+func (ag *collectorClient) GetCollector(context.Context, *connect.Request[collectorv1.GetCollectorRequest]) (*connect.Response[collectorv1.Collector], error) {
 	return nil, nil
 }
-func (ag *agentClient) CreateAgent(context.Context, *connect.Request[agentv1.CreateAgentRequest]) (*connect.Response[agentv1.Agent], error) {
+func (ag *collectorClient) CreateCollector(context.Context, *connect.Request[collectorv1.CreateCollectorRequest]) (*connect.Response[collectorv1.Collector], error) {
 	return nil, nil
 }
-func (ag *agentClient) UpdateAgent(context.Context, *connect.Request[agentv1.UpdateAgentRequest]) (*connect.Response[agentv1.Agent], error) {
+func (ag *collectorClient) UpdateCollector(context.Context, *connect.Request[collectorv1.UpdateCollectorRequest]) (*connect.Response[collectorv1.Collector], error) {
 	return nil, nil
 }
-func (ag *agentClient) DeleteAgent(context.Context, *connect.Request[agentv1.DeleteAgentRequest]) (*connect.Response[agentv1.DeleteAgentResponse], error) {
+func (ag *collectorClient) DeleteCollector(context.Context, *connect.Request[collectorv1.DeleteCollectorRequest]) (*connect.Response[collectorv1.DeleteCollectorResponse], error) {
 	return nil, nil
 }
-func (ag *agentClient) ListAgents(context.Context, *connect.Request[agentv1.ListAgentsRequest]) (*connect.Response[agentv1.Agents], error) {
+func (ag *collectorClient) ListCollectors(context.Context, *connect.Request[collectorv1.ListCollectorsRequest]) (*connect.Response[collectorv1.Collectors], error) {
 	return nil, nil
 }