+++ internal/component/otelcol/receiver/prometheus/prometheus.go
@@ -112,7 +112,7 @@ func (c *Component) Update(newConfig component.Arguments) error {
 		// When supported, this could be added as an arg.
 		trimMetricSuffixes = false
 
-		enableNativeHistograms = true
+		enableNativeHistograms = c.opts.MinStability.Permits(featuregate.StabilityPublicPreview)
 
 		gcInterval = 5 * time.Minute
 	)